A Working Views (DAO)
We might have to create an extension function to convert POCO into DAO
and vice versa in the near future. (if it's possible) It takes too much
time and effort and lines to write these conversion by hands.
On the other hands, the index for all applications work fine. Not sure
of the Applied table though, right now it's possible to add an
application, but applied is unclear.