failed assessments now delete app and applicant
When an applicant applies for a job opening and fails the related
assessment questions, the related applicant and application data is not
supposed to be saved, according to CR48 in the requirements. This
feature is now implemented with this code update.

(Please see documentation screenshots for a visual explanation of how
this works.)